Moved logic for switchinge levels from MapPresenter to Map
