+++ OOPP/app/src/main/java/com/example/sauronsarmy/oopp/Map/Map.java
@@ -174,4 +174,116 @@ class Map implements MapMVPInterface.ModelOps {
                 return -1;
         }
     }
+
+    public boolean tryChangeAreaLevel(int level, int area) {
+
+        if(area ==0){
+            changeArea(area);
+            if(level ==0){
+                changeLvl(level);
+                return true;
+            }
+            else {
+
+                if(getCurrentArea().getLevel(level-1)!=null && (getCurrentArea().getLevel(level-1).getComplete())){
+                    changeLvl(level);
+                    return true;
+                }
+
+            }
+        }
+        else {
+
+            if(getArea(area-1).getComplete() && getAreas().length>area){
+                changeArea(area);
+                if(level ==0){
+                    changeLvl(level);
+                    return true;
+                }
+                else {
+
+                    if(getCurrentArea().getLevel(level-1)!=null && (getCurrentArea().getLevel(level-1).getComplete())){
+                        changeLvl(level);
+                        return true;
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+    private void changeLvl(int index) {
+
+        getCurrentArea().setCurrentLevel(getCurrentArea().getLevels()[index]);
+    }
+    private void changeArea(int index){
+        setCurrentArea(getArea(index));
+        int imgref= getArea(index).getImgRef();
+        setBackgroundRef(imgref);
+    }
+
+
+
+    public void nextLevel(){
+
+        int lvlpos = getLevelIndex() +1;
+        if(lvlpos >= getCurrentArea().getLevels().length){
+            if(getAreaIndex()+1 < getAreas().length){
+
+                tryChangeAreaLevel(0,getAreaIndex()+1);
+            }
+
+        }
+        else{
+            tryChangeAreaLevel(lvlpos,getAreaIndex());
+        }
+    }
+
+    public void previousLevel(){
+        int lvlpos = getLevelIndex() -1;
+        if(lvlpos <0){
+            if(getAreaIndex()-1 >=0){
+
+                tryChangeAreaLevel(getArea(getAreaIndex()-1).getLevels().length -1 ,getAreaIndex()-1);
+            }
+
+        }
+        else{
+            tryChangeAreaLevel(lvlpos,getAreaIndex());
+        }
+
+    }
+
+    private int getLevelIndex(){
+
+        int pos =0;
+        Level curLvl = getCurrentLevel();
+        Level[] lvls = getCurrentArea().getLevels();
+
+        for(int i =0; i< lvls.length ; i++){
+            if(lvls[i].equals(curLvl)){
+                pos =i;
+                break;
+
+            }
+        }
+        return pos;
+    }
+
+
+    private int getAreaIndex(){
+
+        int pos =0;
+        Area currentArea = getCurrentArea();
+        Area[] areas = getAreas();
+
+        for(int i =0; i< areas.length ; i++){
+            if(areas[i].equals(currentArea)){
+                pos =i;
+                break;
+
+            }
+        }
+        return pos;
+    }
 }
